+++ ChangeLog.rst
@@ -10,3 +10,4 @@ ChangeLog
 - Fixed #2: Some routine files such as readme, license, etc.. .
 - Fixed #3: Added more tests.
 - Fixed #4: Added a tiny readme.
+- Fixed #5: Wrote Docstring and Doctests for ``Base64Field`` class.